Making the jquery file locations configurable
Can now be set through config files. Makes it easier for me to switch
to local copies of the files so I can work offline.

I didn't add command-line flags for this stuff because I don't think
I'm likely to use them, but perhaps this indicates that I should have
it set up to automatically generate one from the other?