Improved editing ui! (not wired up)
Gave up on wtforms, they weren't a good fit for the sort of dynamic
thing I wanted to do. Instead I've got a pure javascript
implementation that allows you to create a list of different changes
you'd like to make.

Nothing is hooked up on the backend, yet, and the javascript for this
is pretty miserable. I'm probably going to re-write it. I think now's
the time to learn jQuery.